pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Ruan Oliveira <127878189+RuanHOliveira@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.2"
langchain-core = "^0.3.21"
langchain-ollama = "^0.2.1"
langchain-community = "^0.3.9"
langchain-huggingface = "^0.1.2"
python-dotenv = "^1.0.1"
sentence-transformers = "^3.3.1"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
